Use MAXPATHLEN to avoid path truncation.
This patch modifies all tests in the ftest directory to  use
MAXPATHLEN on all character arrays that store path/file names rather
than arbitrarily sized arrays. This corrects issues where there was
truncation happening when copying a path from one array to another if
the path was long enough. The ftest code may delete the tmp directory
instead of the tmp file due to truncation, and cause LTP to report a
failure/warning when the tmp directory disappears. We saw this in
ftest03,07 where it was trying to copy wdbuf[MAXPATHLEN] into fuss[40]
here:

   sprintf(fuss, "%s/ftest03.%d", getcwd(wdbuf, sizeof( wdbuf)),
getpid());

and then attempting to delete the file but instead got the directory
here:

   execl("/bin/rm", "rm", "-rf", fuss, NULL);

Signed-off-by: Henry Yei <hyei@mvista.com>
Acked-by: Garrett Cooper <yanegomi@gmail.com>