Clean up arguments passed in sighold02.
The functions read_pipe and write_pipe in sighold02.c each expect a
single file descriptor as arguments, however the function prototypes
define no arguments.  Also, the function references to read_pipe pass
two arguments, a file descriptor and '0'.  The second argument should
not be passed, as it is not used.

Attached is a patch to make the function argument definitions and uses
consistent within this testcase.

Signed-off-by: d.marlin <dmarlin@redhat.com>
Acked-by: Garrett Cooper <yanegomi@gmail.com>