Fix at_deny01 in positive and negative cases.
[NOTE: item 3 was not `fixed' because it's an environment issue with the
box]

There are three failures found when running the LTP at_deny01 case.

[1] at_deny01    1  TCONF  :  at command not found on system
[2] /home/test_user_1/at_deny01: line 129: tst_resm: command not found
[3] You are required to change your password immediately (root enforced)

The root causes of these failures:

[1] In the shell script, if the return value of a function is 0, it view
as TRUE. In the at_deny01
162 if type at > /dev/null; then
163         tst_resm TCONF "at command not found on system"
if the system has at function, shell script will print "at command not
found on system". This is opposite.

[2] In the at_deny01, root user will create user named a test_user_1,
copy the file at_deny01 to the home directory of test_user_1,
change user to test_user_1 and run the at_deny01 script.
when run the script, ${LTPROOT}/testcase/bin is in the environment
variable of root, but NOT in the test_user_1. So when the test_user_1
run this script, it can NOT find the tst_resm command.

[3] When some boards boot up, the system time is 010100001970. the Linux
system will let NON-root user change its password immediately
and stop the script running. This patch will let the script run normally
without changing the password.

Signed-off-by: Garrett Cooper <yanegomi@gmail.com>