Fix testcase through simplification.
As reported by Wei Yang from Windriver, there is an issue with the testcase
where it's failing with EINVAL on select platforms.

In order to fix the problem, a lot of unnecessary complexity needed to
be removed from the testcase with setup, pointer arithmetic -> array
traversal, etc.

So, the test is a lot more linear and easier to follow now. More work
may need to be done to get the other IPC testcases to conform to a
similar paradigm (there's a fair amount of code sharing between the
testcases).

Signed-off-by: Garrett Cooper <yanegomi@gmail.com>