Don't assume we know the value of __NR_cacheflush.
MIPS on Linux is a complicated architecture. There's o32, n32, and n64,
and a ton of different variants playing on those 3 major players.

Each syscall set has a different set of numbers, so the prior assumption
was wrong (__NR_Linux + 147), as pointed out by Feno Gao
(feng.gao@windriver.com). So if the syscall doesn't exist, don't assume
that the value is __NR_Linux + 147, and just set the value to 0 (which
will make the outcome inconclusive as we don't honestly know if we _are_
testing what we think we are testing).

Signed-off-by: Garrett Cooper <yanegomi@gmail.com>