Use time_t instead of unsigned long / int.
The test was using a mish-mash of unsigned long and int types to pass
data between the thread and evaluate values, so the problem lied in
potential truncation of data.

time_t is a 32-bit type or 64-bit type on some OSes, so it's best to
remain consistent with a single fixed data type instead of moving
between int (which can be 32-bit or 64-bit, depending on the OS and
architecture) and unsigned long (which can be 32-bit or 64-bit,
depending on the OS and architecture).

Acked-by: Gui Jianfeng <guijianfeng@cn.fujitsu.com>
Signed-off-by: Garrett Cooper <yanegomi@gmail.com>