Fix aio_return conformance testcases.
The problem with the testcases is that the original writer assumed that
the values were stored in the global errno instead of a value
retrievable via aio_error, and ignored the following item in the POSIX
spec that states:

If the error status for the operation is equal to [EINPROGRESS], then
the return status for the operation is undefined.

Thus, the results were incorrect in some cases, as determined through
testing by Mitani-san.

Signed-off-by: Garrett Cooper <yanegomi@gmail.com>