Fix alignment on strict archs (arm, sh, etc).
For all the architectures that have stack that grows downwards, the
stack argument passed to the clone function points to the topmost
address of the memory space set up for the child stack. So it is not
correct to move the stack down. The ABI guarantees that the function has
to create it's own stack frame.

Nevertheless, on some architectures that do not accept unaligned stack,
it causes the process to be killed by the kernel with a SIGBUS.
On architectures that safely handle unaligned stack, we are actually
reserving one byte less.

Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>