Fix fork/7-1 test:
The fork 7-1 testcase tests that child gets a copy of the parent message
catalog descriptor. For that the test creates catalog from file
messcat_src.txt that is not being installed now and even if would be,
the test has wrong hardcoded path. Fix it by making the test self-contained
so it creates and fills the input file, generates message catalog, opens
it, forks, etc ...

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>