Add pid reuse race condition test.
A program that repeatedly forks and waits is susceptible to having the
same pid repeated, especially when it competes with another instance of
the same program.  This is really bad for bash implementation.
Furthermore, many shell scripts assume that pid numbers will not be used
for some length of time.

Race Description:

A                                B

// pid == offset == n            // pid == offset == n + 1
test_and_set_bit(offset, map->page)
                                test_and_set_bit(offset, map->page);
                                pid_ns->last_pid = pid;
pid_ns->last_pid = pid;
                                // pid == n + 1 is freed (wait())

                                // Next fork()...
                                last = pid_ns->last_pid; // == n
                                pid = last + 1;

Signed-off-by: CAI Qian <caiqian@redhat.com>