This kind of reverts commit: 9ab64b8c3151b439fda4fd8ca0a8f6f012cd7458
The io_contex_t is defined as a pointer to io_context in libaio.h, which
is passed to syscall and filled with user_id in kernel. The type of
user_id in kernel sure is unsigned long, but the libaio expects that
they work with struct io_context *.

Anyway we go back to the io_context_t as the test won't compile with
aio_context_t. The problem is that libaio is broken for some time and
nobody noticed, I'll contact libaio developers about it.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>