Deal with inconsistent sender names in mail0[23].
mail02 and mail03 tests non-existed host and non-existed
username in the test. However in some systems (at least in RHEL6),
they both fails because of the sender name of the returned letter is
"Mail Deliver System" instead of "Mail Delivery Subsys", which makes RC1
to be zero and fail the test. The patch fixes this issue.

Signed-off-by: Caspar Zhang <czhang@redhat.com>