Check legacy and non-legacy cpuset paths in test.
From

http://lxr.linux.no/#linux+v2.6.37/Documentation/cgroups/cpusets.txt

"Note that for legacy reasons, the "cpuset" filesystem exists as a
wrapper around the cgroup filesystem.

The command

mount -t cpuset X /dev/cpuset

is equivalent to

mount -t cgroup -ocpuset,noprefix X /dev/cpuset"

However, I found this is not always case. In addition, there is an
direct implementation without cgroup, and I can't find any doc about
the behaviours of the prefix.

So, It can be difficult to predict if cpuset files have prefix or not,
so only to fail after checked both.

Signed-off-by: CAI Qian <caiqian@redhat.com>
Signed-off-by: Han Pingtian <phan@redhat.com>