Fix clock_getcpuclockid test.
The test was expecting that two clockid_t poiting to the same clock must
be the same, which is not required by POSIX. Now the tests, if ids are
different, check if reported time is aproximetly the same.

This also introduces timespec_nsec_diff() function, that should be used
to get nanosecond differerence of two struct timespec values.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>