add config.h header file in mmap10 test
Hi all, mmap10 test will check the macro "HAVE_MADV_MERGEABLE" to see if
the kernel supports MADV_MERGEABLE argument which is available since
2.6.32, however in mmap10.c it didn't include the header file "config.h"
which may cause "HAVE_MADV_MERGEABLE" undefined, so that the test will
never be executed correctly. This patch is simple enough but do fixing
the issue.

Signed-off-by: Caspar Zhang <czhang@redhat.com>