Complete fix noted previously committed for acct*
A follow-up fix for http://article.gmane.org/gmane.linux.ltp/13576 .
In fact, it's not a kernel bug of ENOTDIR's failure. ENOTDIR means that
you setup a string as directory in argument, but it is actually a file.
The patch fixes this failure. The result of given path
/tmp/does/not/exist should be expected as ENOENT.

And also, there is a mistake in EPERM case, to restore the permissions,
SAFE_SETUID should be used instead of SAFE_SETGID, else it will never
recover the permissions.

Signed-off-by: Garrett Cooper <yanegomi@gmail.com>