Add mlock DoA testcase for kernel regression.
This is a reproducer copied from one of LKML patch submission,
which subject is

[PATCH] mlock: revert the optimization for dirtying pages and
triggering writeback.

"In 5ecfda0, we do some optimization in mlock, but it causes
a very basic test case(attached below) of mlock to fail. So
this patch revert it with some tiny modification so that it
apply successfully with the lastest 38-rc2 kernel."

Signed-off-by: Han Pingtian <phan@redhat.com>