mm: mtest01(w): fail to get EOF on some archs
on some secondary archs (e.g. ppc64), mtest01(w) tests fails for
incorrect argument. Digged into the problem, we found this sentense:
  while ((c = getopt(argc, argv, "c:b:p:wvh")) != -1)
here 'c' is char type instead of int. It runs well on x86 systems, but
gets failed on ppc64, s/390x systems, since it returns 255 instead of -1
when 'c' is char type.
This failure was introduced in commit
8147b761b08eefde4f9b965fabb9ef614b3d5817, this patch fixes the issue.

Signed-off-by: Caspar Zhang <czhang@redhat.com>