ltp: move type definitions up in "doio.c"
The definitions of some types are spread around in the file
"testcases/kernel/fs/doio/doio.c".  Group them near the top of the
file.  This is in preparation for the next change, which adds
full-fledged C prototypes for all functions in the file.

All changes in this patch are simple "bulk moves" of blocks of
code from one place in the file to another.

Signed-off-by: Alex Elder <aelder@sgi.com>