ltp: use C function prototypes in "doio.c"
Many functions in "doio.c" are defined using old-style C
declarations.  Convert all functions so they follow standard C,
and include standard prototypes for each.

Every function (except main()) now has a standard C prototype, and
all prototypes are grouped in the same section of the file.  The
prototypes--as well as the function definitions themselves--now
follow a consistent formatting style.

One more change was made, to some signal handler functions.
Previously, a few of them were defined without an integer argument,
and so did not comply with the definition of a sighandler_t function
(defined in signal(2)).  These have been fixed to conform.

Signed-off-by: Alex Elder <aelder@sgi.com>