ltp: fix warnings in "growfiles.c"
In "testcases/kernel/fs/doio/growfiles.c" there are a number of
variables of type off_t which are formatted using a "%ld" conversion
specification.  When I build on an i386 system I get warnings,
because evidently off_t is not defined as a long in this
environment.

Updated to address Garrett Cooper's concerns that variables named
"tmp" are used for a number of purposes in this file.  In one case
I renamed the variable "off_tmp" (its type is off_t) (but did not
change corresponding printed text).  The other cases the "tmp"
variable was more appropriately type int.  Also adjusted some
whitespace, as requested.

Signed-off-by: Alex Elder <aelder@sgi.com>