ltp: fix warnings in "fsstress.c"
Fix remaining warnings in "testcases/kernel/fs/fsstress/fsstress.c".
Three of them are due to mismatches between fprintf() conversion
specifications and the variables being formatted.  Fix these by
casting the variables to the right type.

Two others are due to the use of a __uint64_t local variable to
represent a value of type __u64 (which evidently can be defined
differently in some environments).  Fix these by using the exact
matching type in the cases that produce the warning messages.

Updated to address the concern raised by Garrett Cooper about the
use of types that are not identical to what's used in the XFS
ioctl() interface in the kernel.  Also added memset() calls so we're
sure no garbage resides in data structures passed between user space
and the kernel.  Also updated some whitespace, as requested.

Signed-off-by: Alex Elder <aelder@sgi.com>