Remove memctl test 4
The test relies on the exact value of memory.usage_in_bytes but the internal
kernel accounting for this counter has changed with
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=cdec2e4265dfa09490601b00aeabd8a8d4af30f0
(merged in 2.6.33-rc1) upstream commit which implemented per-cpu batch
charging to improve scalability.

This means that the value of memory.usage_in_bytes is updated in
CHARGE_BATCH (currently 32 pages) chunks so the exported value may be in
+-NUR_ONLINE_CPUS*CHARGE_BATCH range.

There is also another optimization for uncharging which is done in batches
as well. This, however, is not per-cpu but rather per task to make exit (or
more precisely memory regions unmap) path as fast as possible. Very roughly
it marks a block of uncharges which are counted per task and once the block
is done the accumulated value is subtracted from the resource counter.

This means that the value is rather fuzzy and we cannot use it for any exact
comparisons.

Signed-off-by: Michal Hocko <mhocko@suse.cz>