ftrace: fix memory allocation size computation
The issue with ftrace_buffer_size.sh is that it allocates 10% of
server's memory. However the tracing ring buffer is implemented using
separate buffers for each CPU in the system. And putting new value to
the buffer_size_kb file increases each CPU buffer with new value. Now
having system with more than 10 CPUs we will get out of memory sooner or
later.

The patch makes sure the memory size is divided by number of CPUs in the
system.

Signed-off-by: Jiri Olsa <jolsa@redhat.com>
Signed-off-by: Caspar Zhang <czhang@redhat.com>