Rewrite the policy test.
The current policy test (1-1.c) is inadquate and returns a false positive.
The deficiencies of the test are:

o fails to set the inheritance policy within the attributes, so the
  policy setting is completely ignored during thread creation.
o fails to set a priority in the attributes.
o fails to verify the created thread's policy and priority.
o only attempted to test the SCHED_FIFO policy.

Fix this test by creating 3 new test src files to test SCHED_FIFO,
SCHED_RR, and SCHED_OTHER. Since the bulk of the test code is identical,
place the bulk of the code into a common file that is included in the three
individual test cases.

The tests for SCHED_FIFO and SCHED_RR also set and verify a priority.

Also normalize the error messages as well as make them more informative.

Signed-off-by: Peter W. Morreale <pmorreale@novell.com>