Fix getschedparam 1-3 test.
This test fails when attempting to set the priority in an thread
attributes because of out-of-order operations on setting the attributes.

The sched policy must be set first in the attributes, then you can set a
prirotiy.  This is because the default policy is SCHED_OTHER, which does
not have a priority associated with it.  So move the setting of the
policy before attempting to set the priority.

Signed-off-by: Peter W. Morreale <pmorreale@novell.com>