Redo the 4-1 test.
The current 4-1 test attempts to see if EINTR can be returned from a
pthread_getschedparam(), however it replies on a non-deterministic
algorithm.  It creates a thread and sends signals in a tight loop, and
loops doing a pthread_getschedparam() in another thread for a
time period of 1 second, attempting to see if EINTR is returned.

We can fix that by creating a single thread which spins in an infinite
loop waiting until a signal is posted by a searate thread.

Because we know that the signal was posted prior to the call, we can
guarantee the knowledge of the return code.

This patch replaces the current 4-1.c source code.

Signed-off-by: Peter W. Morreale <pmorreale@novell.com>