Replace wait() with waitpid() in fork13.c
I got the following FAIL, and this FAIL often happens when running the
fork13 test with -C option.  See this:

./fork13 -c 3 -I 1
fork13      1  TPASS  :  5729 pids forked, all passed
fork13      1  TFAIL  :  Wait return value: expected pid=10511, got 32132, iteration 2957.
fork13      1  TPASS  :  2755 pids forked, all passed

That is because the first process is the father of the other two
processes, and when the father calls wait(), any of the children could
be waited.  So I think we should use waitpid() instead of wait().

Acked-by: Caspar Zhang <czhang@redhat.com>
Signed-off-by: Tang Chen <tangchen@cn.fujitsu.com>