Remove 1-1 pthread_mutex_getprioceiling test
The 1-1 test is in error since the test attempts to return the priority
ceiling for a mutex using the PTHREAD_PRIOR_INHERIT protocol.

Priority ceilings are not defined for the INHERIT protocol under a SCHED_OTHER
scheduling class.  They are defined for the PTHREAD_PRIO_PROTECT and are
limited to the range of SCHED_FIFO priorities.  See section 13.6.1.2 of the
POSIX standard.

Note that the 1-2 test sets the PTHREAD_PRIO_PROTECT protocol and is
otherwise identical to the 1-1 test.

So, remove the 1-1 test and rename the 1-2 test to become the new 1-1 test.

Signed-off-by: Peter W. Morreale <pmorreale@novell.com>