Redo the sigaction assertion 16 tests.
The suite of 16-* tests for sigaction individually test SA_RESTART for a
specific signal.  The tests worked by creating a handler for the signal,
then creating a thread which blocks in a sem_wait().  The main thread
sends a signal to the child, and ensures that the child thread
remains in sem_wait() (obstentively performing a restart of the sys
call).

This patch replaces all the individual test files with a single program
that loops through testing each signal using the same methodology.

For each signal, two sigactions are performed.  One to set a handler for
each signal tested, then another to reset the action back to the
default.

In addition, the old test attempted to have the main thread wait for the
child thread to enter sem_wait() by performing 3 sched_yield()s.  This
is likely the cause of the regressions as there was no guarantee that
the child actually executed to the point of entering sem_wait() prior to
receipt of the signal.

The new algorithm does this differently.  This patch now bumps the sched
policy and priority into the SCHED_FIFO class (priority 10) and uses a
predicate between the child and main threads.  While not a guarantee,
the likelihood of the child being preempted during this race is much
lower.

Signed-off-by: Peter W. Morreale <pmorreale@novell.com>