kill05: Added tst_resm() to print the proper error code(TPASS) if test exits with EPERM
Added "tst_resm()" call to print the proper LTP error code(TPASS)
in case of the testcase exit with expected result.

The testcase didn't call "tst_resm()" in case of
exit without error i.e. after that EPERM is correctly raised.
The proper LTP error code, in that case, is TPASS as EPERM
is the expected result.

Signed-off-by: Francesco Rundo <francesco.rundo@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>