posix:clock_getcpuclockid:6-1.c:change the argument -1 to -2
If -1 is passed to the argument pid of clock_getcpuclockid(),
the macro MAKE_PROCESS_CPUCLOCK(pid,CPUCLOCK_SCHED) uses it
to make the pidclock to 2, and than passes the number 0x2
to the system call clock_getres().

The posix says that if _POSIX_CPUTIME is defined, implementations
shall support clock ID values obtained by invoking clock_getcpuclockid(),
which represent the CPU-time clock of a given process, implementations
shall also support the special clockid_t value CLOCK_PROCESS_CPUTIME_ID,
which represents the CPU-time clock of the calling process when invoking
one of the clock_*() or timer_*() functions.

So passing -1 as the argument pid represents the special CLOCK_PROCESS_CPUTIME_ID.

This patch change -1 to -2 as the argument. Make the test to a normal one, but
not a special used value.

And acctually, the Linux kernel has already supported this special one.

Signed-off-by: Wanlong Gao <gaowanlong@cn.fujitsu.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>