mm/vma01: new testcase
vma01 - test not merging a VMA which cloned from parent process

This program is used for testing the following upstream commit:
965f55dea0e331152fa53941a51e4e16f9f06fae

The cloned VMA shares the anon_vma lock with the parent process's
VMA. If we do the merge, more vmas (even the new range is only
for current process) use the perent process's anon_vma lock. This
introduces scalability issues. find_mergeable_anon_vma() already
considers this case.

This test program clones VMA and checks /proc/$pid/maps file, on
an unpatched kernel, there is a single 6*ps VMA for the child
like this:

7fee32989000-7fee3298f000 -w-p 00000000 00:00 0

On a patched kernel, there are two 3*ps VMAs like this:

7f55bbd47000-7f55bbd4a000 -w-p 00000000 00:00 0
7f55bbd4a000-7f55bbd4d000 -w-p 00000000 00:00 0

Signed-off-by: Caspar Zhang <czhang@redhat.com>