sched_yield 1-1 fix
The current sched_yield 1-1 test floods available processors with
SCHED_FIFO processes (nr_cpus - 1), then starts two threads that verify
the behavior of sched_yield and affinitizes these threads to the remaining
processor.

This is silly and not useful if we assume that setting affinity works
correctly (which this type of test depends upon)

So rewrite the test to merely use a single processor.  We test for success
by starting a child that loops doing sched_yield()s and then have the
parent kill it.  The parent can only kill it if sched_yield() works. (ok,
and affinity too...)

[Note: this patch moves the non portable affinity code into common
       header that needs to be ported to your favorite UNIX]

Signed-off-by: Peter W. Morreale <pmorreale@novell.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>