Fix sigqueue 9-1 test
Two issues with the last patch.

o Add string.h to the list of includes for strerror()

o Current test does not handle rt signals delivered while the test is
running.  This screws up the pending count and causes a false failure.

Fix the second issue by setuid() to an unused uid.

[The problem was well analyzed by:
 Wanlong Gao <gaowanlong@cn.fujitsu.com>]

Signed-off-by: Peter W. Morreale <pmorreale@novell.com>