Fix timer_gettime 1-3
The test sets a time and sleeps for a bit, then queries the remaining time
and applies a delta to complete the test.

Most likely, the issue with this test is merely one of scale.  The test
is measuring nanoseconds and likely the times are so low that scheduling
delays (xen), NMIs/etc can effect success of failure.

So increase the times a little bit to overcome these delays.

Signed-off-by: Peter W. Morreale <pmorreale@novell.com>