rt_sigqueueinfo01: Fix test failed on mips.
rt_sigqueueinfo01 case test failed on mips arch:
	rt_sigqueueinfo01    1  TFAIL  :  Test Failed
	rt_sigqueueinfo01    2  TFAIL  :  Test Failed

We are sending signal 17 that could be SIGUSR2, SIGCHILD or SIGSTOP and depends on hardware platform.
On x86, arm, and powerpc, this number is defined to SIGCHILD. So in this case I think it should be SIGCHILD.
However, on mips this number is defined to SIGUSR2 and SIGCHILD is defined as 18.
This made the test case failed. I think that symbolic name rather than number should be used.

And this patch also add a a missing brace.

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>