pthread_create:remove the pointless testcase
In posix Issue 7: the informative sections of pthread_create said that:
The [EINVAL] error for an uninitialized thread attributes object is removed;
this condition results in undefined behavior.

So, the test for uninitialized thread attributes objiect is pointless,
it no need to test an undefined behavior. Just remove this case.

Signed-off-by: Wanlong Gao <gaowanlong@cn.fujitsu.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>