runtest/fs: change parameters of growfiles
gf16/gf17 create 10 files and keep them growing and shrinking.
But when the number of created file (-S) can be divided exactly
by the rate shrinkfile() is involked (-T), the shrinking will
always happen on the same file. This may cause the tests run out
of memory and unexpected OOM killer be launched.

The amount of consumed memory depends on the speed of CPU/Bus.
In our machine, CPU 400 MHz/Bus 100 MHz, about 240 MB free memory
was consumed before the test finished.

To prevent this, we adjusts the shrinkfile() rate (-T) slightly
to ensure it cannot be divided by the number of created file (-S).

Signed-off-by: Andrew Chen <ypchen@faraday-tech.com>