Second fix for sigaction/16-1.c
If we failed to set the thread priority in the second thread we should
rather call exit() than trying to propagate the error value, which would
lead to several types of side effect (one of them is possibility of
blocking the original thread in infinite loop).

Also return UNTESTED on EPERM which is sematically better.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>