aio_fsync: merge 4-1 and 4-2
As Cyril said, seeing the difference between 4-1.c and 4-2.c these two should be
merged into one testcase.

In the event that any of the queued I/O operations fail, aio_fsync()
shall return the error condition defined for read() and write().
The error is returned in the error status for the asynchronous fsync()
operation, which can be retrieved using aio_error().

Correct test would check the return value from aio_error(), which is
defined for read() and write().
Mind that you could call them it exactly once after the
operation is finished.

Signed-off-by: Wanlong Gao <gaowanlong@cn.fujitsu.com>
Signed-off-by: lidan <li.dan@cn.fujitsu.com>