readlink03: fix return value for the syscall readlink()
Since commit http://git.kernel.org/linus/65cfc672
which added O_PATH support to the *at() calls,
the error return for readlink/readlinkat
for the empty pathname has switched from
ENOENT to EINVAL. And this patch merged to linux v2.6.39,
So LTP commit https://github.com/linux-test-project/ltp/commit/0cc7514
fix this with the kernel version v2.6.39 compared.

But recently, commit http://git.kernel.org/linus/1fa1e7f6
which ensure we return ENOENT for the empty pathname
for normal lookups.

Now, the patch whitch switch the return value back again
is already queued for 3.0 and 3.1 stable release.

So, if the return value is EINVAL for empty pathname,
it may be a kernel bug, so we needn't to check the kernel
version but return the "kernel bug" warning  to LTP the users.

Signed-off-by: Wanlong Gao <gaowanlong@cn.fujitsu.com>
Acked-by: Jan Stancek <jstancek@redhat.com>
Acked-by: Caspar Zhang <caspar@casparzhang.com>