ext4-online-defrag: use the e4defrag from local system
Remove the out-of-dated e4defrag prog, and use the e4defrag
installed from e2fsprogs package.

Note:As Kazuya said, e4defrag here is too old to defrag the
file in the latest kernel. We should use the latest e4defrag
command that is contained in e2fsprogs. So, when testing,
install the lastest e2fsprogs by self.

Signed-off-by: Wanlong Gao <gaowanlong@cn.fujitsu.com>
Tested-by: Kazuya Mio <k-mio@sx.jp.nec.com>