ptrace: check signal status in main thread
Newer versions of gcc will assume "child_stopped" is not set async (via
the signal handler) and so won't reload the value leading to an infinite
loop.  Rather than marking the variable as volatile, just move the signal
checking into the main thread.  This avoids cpu burning and should work
the same.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>