syscalls/sendfile08: new test case for sendfile(2) syscall
A bug in the splice code had caused the file position on the write side
of the sendfile system call to be incorrectly set to the read side file
position. This can result in the data being written to an incorrect offset.

This is a regression test for kernel commit
2cb4b05e7647891b46b91c07c9a60304803d1688

Signed-off-by: Eryu Guan <eguan@redhat.com>