mm/vma03: new testcase
vma03 - CVE-2011-2496 avoid wrapping vm_pgoff in mremap()

This is a reproducer for CVE-2011-2496.

The normal mmap paths all avoid creating a mapping where the pgoff
inside the mapping could wrap around due to overflow.  However, an
expanding mremap() can take such a non-wrapping mapping and make it
bigger and cause a wrapping condition. There is also another case
where we expand mappings hiding in plain sight: the automatic stack
expansion.

This program tries to remap a mapping with a new size that would
wrap pgoff. Notice that it only works on 32-bit arch for now.

Signed-off-by: Caspar Zhang <caspar@casparzhang.com>