Fix runtest file for fork13.
The '-c 2' parameter causes the test to fork at
the test very beginning, however the information that
two copies of the test exist is not propagated to the
ltp-pan.

So either the return value from the second test
instance is completely ignored, or the second
test instance is killed once the first instance
exits which fills the logs with bogus test failures.

This patch removes the '-c 2' from syscalls runtest
file, however as the '-c' option is broken by design,
it may be removed completely in the future.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>