mem: Fix static linking.
The libmem now uses safe_strtol() that is implemented in libltp, which
causes build problem as the -lltp linker flag is passed always first,
the safe_strtol() is not feetched from the library which is needed by
the libmem. Fix that by adding the -lltp flag second time in particular
Makefiles.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>