creat07: Drop basename() function call from read-only variable
test_app is a read-only variable because it is defined as a preprocessor macro.
Using basename() on this variable, it will most certainly lead to a segmentation
fault in uClibc code. Moreover, there is no need to use basename() at all since
the test_app variable is a single string and it does not contain any leading
paths.

Quote from uClibc's include/libgen.h header:

"Return final component of PATH.
	This is the weird XPG version of this function.  It sometimes will
	modify its argument.  Therefore we normally use the GNU version (in
	<string.h>) and only if this header is included make the XPG
	version available under the real name.

	extern char *__xpg_basename (char *__path) __THROW;
	#define basename        __xpg_basename"

Signed-off-by: Markos Chandras <markos.chandras@imgtec.com>