mm/cpuset01: make mem_hog() get executed
cpuset01 will run for a period of time which is defined via -I option,
during that period, three things will be done:

    1) one thread loops to change memory nodes;
    2) one thread loops to wait its children doing memhog, then sends
       SIGUSR1 to make them exit;
    3) a bunch of children doing memhog()

However this is not true for 2) in this case, the thread doesn't loop
for -I <time> to wait its children to finish. The fix should be simple
enough by adding a for-loop with TEST_LOOPING here.

Signed-off-by: Caspar Zhang <caspar@casparzhang.com>
Acked-by: Zhouping Liu <zliu@redhat.com>